requirements.txt
alembic==1.7.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==3.2
importlib-resources==5.2.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
Mako==1.1.5
MarkupSafe==2.0.1
numpy==1.19.0
pandas==1.0.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
scikit-learn==0.24.2
scipy==1.5.0
six==1.16.0
sklearn==0.0
sklearn-pandas==2.2.0
SQLAlchemy==1.4.23
threadpoolctl==2.2.0
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
psycopg2==2.7.7
